wip: optimizations
